Struct booklibrs::books::book::Book

pub struct Book {
    pub(crate) the_book: Option<Weak<RefCell<TheBook>>>,
    pub(crate) is_using: bool,
    pub(crate) cabinet: u16,
    pub(crate) shelf: u8,
    pub(crate) readers: Vec<(Weak<RefCell<Reader>>, (Date, Date))>,
}

Simple Book structure, which contains title, author, amount of pages, using status, location (number of cabinet and it's shelf) and readers with start and finish dates

Fields
the_book: Option<Weak<RefCell<TheBook>>>is_using: boolcabinet: u16shelf: u8readers: Vec<(Weak<RefCell<Reader>>, (Date, Date))>
Implementations

impl Book

pub(crate) fn new(
the_book: Rc<RefCell<TheBook>>,
app: &App,
lang: Lang
) -> Option<Self>

Constructs new simple book. By default it has no readers and it isn't used. If there are some input errors, it'll return None

pub(crate) fn restore(
_the_book: Rc<RefCell<TheBook>>,
_is_using: bool,
_cabinet: u16,
_shelf: u8
) -> Self

Constructs book with known params. It uses for load and restore

pub(crate) fn title(&self) -> String

Gets title

pub(crate) fn author(&self) -> String

Gets author

pub(crate) fn pages(&self) -> u16

Gets amount of pages

pub(crate) fn find_reader_first(
&self,
reader: &Rc<RefCell<Reader>>
) -> Option<usize>

Searches reader. If it isn't found, it' ll return the amount of all readers. else it will return reader index of the first occurrence

pub(crate) fn find_reader_last(
&self,
reader: &Rc<RefCell<Reader>>
) -> Option<usize>

Searches reader. If it isn't found, it' ll return the amount of all readers. else it will return reader index of the last occurrence

pub(crate) fn remove_reader(&mut self, reader: &Reader) -> &mut Self

Removes reader

pub(crate) fn remove_all_readers(&mut self) -> &mut Self

Removes all readers of book

pub(crate) fn start_reading(
&mut self,
reader: &Rc<RefCell<Reader>>,
date: Date
) -> &mut Self

Function that uses after giving book to reader. It adds reader (converts rc to weak), start and return dates.

pub(crate) fn finish_reading(&mut self) -> Result<&'_ mut Self, u8>

Function that uses after returning book from reader. It changes book's status and finish date

pub(crate) fn change_location(
&mut self,
new_cabinet: u16,
new_shelf: u8
) -> &mut Self

Changes cabinet's and shelf's number where book is located.

pub(crate) fn clone(&self, the_book: Rc<RefCell<TheBook>>) -> Self

Clones simple book with new readers' pointers

pub(crate) fn to_string(&self, book_system: &BookSystem) -> String

Represent Book as String with next format:

{title} {author} {amount of pages} ({order number})